// src/types/informePersonalidad.ts
export interface InformePersonalidad {
  estudiante_nombre: string;
  curso_nombre: string;
  informe_id: number;
  estudiante_id: number;
  anio: number;
  formacion_etica_1: string;
  formacion_etica_2: string;
  formacion_etica_3: string;
  formacion_etica_4: string;
  formacion_etica_5: string;
  formacion_etica_6: string;
  formacion_etica_7: string;
  formacion_etica_8: string;
  crecimiento_1: string;
  crecimiento_2: string;
  crecimiento_3: string;
  crecimiento_4: string;
  crecimiento_5: string;
  entorno_1: string;
  entorno_2: string;
  entorno_3: string;
  entorno_4: string;
  entorno_5: string;
  entorno_6: string;
  entorno_7: string;
  aprendizaje_1: string;
  aprendizaje_2: string;
  aprendizaje_3: string;
  aprendizaje_4: string;
  aprendizaje_5: string;
  aprendizaje_6: string;
  aprendizaje_7: string;
  conductas_1: string;
  conductas_2: string;
  conductas_3: string;
  conductas_4: string;
  conductas_5: string;
  conductas_6: string;
  conductas_7: string;
  observaciones: string | null;
  estado: string;
  fecha_creacion: string;
  fecha_actualizacion: string;
} 